requirements.txt
# LangGraph 工作流 + LangChain
langgraph>=0.2.28
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0

# FastAPI 服务层
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# 工具库
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
openai>=1.50.0
